+++ app/src/models/collectionCard.ts
@@ -1,8 +1,6 @@
-// import { CollectionCardData } from "app/types/Collection";
 import { imageURL } from "../utils/utils";
 import { parseBoolean } from "../utils/utils";
 
-// TODO: Connect to typescript interface for CollectionCardData
 export class CollectionCardModel {
   uuid: string;
   title: string;