+++ app/src/utils/utils.ts
@@ -1,16 +1,13 @@
 import {
   ADOBE_ANALYTICS_SITE_SECTION,
   ADOBE_ANALYTICS_DC_PREFIX,
-  ALLOWED_FILTERS,
 } from "../config/constants";
 import CollectionDataType from "@/src/types/CollectionDataType";
 import ItemDataType from "@/src/types/ItemDataType";
-import {
-  AvailableFilter,
-  AvailableFilterOption,
-} from "../types/AvailableFilterType";
+import { AvailableFilterOption } from "../types/AvailableFilterType";
+import type { SearchResultRecordType } from "../types/SearchCardType";
 import type { Highlight } from "../types/HighlightType";
-import { isValidFilter } from "./searchManager";
+import { isValidFilter } from "./searchManager/searchManager";
 
 /**
  * Represents a IIIF Image API URL, which will be used globally throughout the application.
@@ -91,10 +88,6 @@ export const stringToSlug = (string: string = ""): string => {
     .replace(/-+/g, "-"); // remove consecutive hyphens
 };
 
-export const parseBoolean = (value: string): boolean => {
-  return value == "true" ? true : false;
-};
-
 export const titleToDCParam = (string: string = ""): string => {
   return string?.replace(/\s+/g, "+"); // replace spaces with +
 };
@@ -141,6 +134,12 @@ export function formatHighlightText(highlights): Highlight[] {
 
 export const capitalize = (text: string): string => {
   return text?.charAt(0) ? text.charAt(0).toUpperCase() + text.slice(1) : text;
+};
+
+export const getRecordType = (type: any): SearchResultRecordType => {
+  // capitalize first character in strin
+  //  capitalize(type)  // returns type error
+  return type.charAt(0).toUpperCase() + type.slice(1);
 };
 
 export const getCollectionFilterFromUUID = (
@@ -252,8 +251,8 @@ export function highlightTitleWords(title: string, highlights): string {
 }
 
 /* Helper for highlighting search result title given that
-   full title appears in the highlight field. 
-   
+   full title appears in the highlight field.
+
 export const getTitleWithHighlights = (highlights, title) => {
   const titleHighlight = highlights.find(
     (highlight) => highlight.field === "Title"
@@ -270,4 +269,11 @@ export const isDCUuid = (identifier) => {
 
 export const deSlugify = (slug: string): string => {
   return slug.replace(/-/g, " ").replace(/\b\w/g, (c) => c.toUpperCase());
+};
+
+export const parseBoolean = (value: any): boolean => {
+  if (typeof value === "boolean") {
+    return value;
+  }
+  return value == "true" ? true : false;
 };