+++ public/CHANGELOG.md
@@ -7,6 +7,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Added Universal Viewer support and metadata fields
+- Conditionally render viewer with rights logic (DR-3645)
+- Added Playwright testing library (DR-3680)
+- Added Playwright tests for navigation menu, what is public domain and search bar, explore further for the home page (DR-3658)
+- Added Playwright test to search for a keyword (DR-3659)
+
+## [0.4.2] 2025-05-22
+
+### Added
+
+- Added sequence sort option and new default sort functionality to search (DR-3609)
+- Connected collection landing page (metadata) to API (DR-3362)
+- Added collection structure component (DR-3545)
+- Added deprecated division slug redirects
+- Added clientside subcollection redirects (DR-3616)
+
+### Updated
+
+- Updated collection and subcollection filter value format (DR-3515)
+
+## [0.4.1] 2025-05-16
+
+### Updated
+
+- HOT FIX: updated search results to use full title instead of just highlights
+
+## [0.4.0] 2025-05-15
+
+### Fixed
+
+- Fixed filters to be encoded as URI components (DR-3564)
+- Fixed dropdown filters to only display selection (DR-3598)
+
+### Added
+
+- Logged 404 path to New Relic (DR-3550)
+
 ## [0.3.6] 2025-04-21
 
 ### Updated
@@ -20,7 +58,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added search query redirects (DR-3529)
 - Added collection slug to uuid redirects (DR-3538)
 - Connected search page to API (DR-3416)
-- Log path in notFoundPage.tsx to New Relic. (DR-3550)
 
 ## [0.3.5] 2025-03-27